requirements.txt
biopython
pandas
PyBioMed
numpy
matplotlib
plotly
seaborn
scikit-learn
shap
streamlit
pytest
